[CL-2553] Fix the rest of the e2e tests
