+++ front/cypress/e2e/project_topics.cy.ts
@@ -34,12 +34,6 @@ describe('Project topics', () => {
       // Add custom topic
       cy.get('#e2e-add-custom-topic-button').click();
       cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
       cy.get('.e2e-admin-list-row').contains(topicTitle);
@@ -55,12 +49,6 @@ describe('Project topics', () => {
       // create a topic
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
       cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
@@ -96,12 +84,6 @@ describe('Project topics', () => {
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
       cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
       cy.get('.e2e-admin-list-row').contains(topicTitle);
@@ -118,12 +100,6 @@ describe('Project topics', () => {
         .first()
         .find('#e2e-custom-topic-edit-button')
         .click();
-      cy.get('#title_multiloc').clear().type(editedTopicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').clear().type(editedTopicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').clear().type(editedTopicTitle);
-      cy.get('.fr-BE').click();
       cy.get('#title_multiloc').clear().type(editedTopicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
@@ -144,12 +120,6 @@ describe('Project topics', () => {
       // create a topic
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
       cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
@@ -179,12 +149,6 @@ describe('Project topics', () => {
       // create a topic
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
       cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
@@ -238,12 +202,6 @@ describe('Project topics', () => {
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
       cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
 
@@ -273,12 +231,6 @@ describe('Project topics', () => {
       // create a topic in the topic manager
       cy.visit('admin/settings/topics');
       cy.get('#e2e-add-custom-topic-button').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-BE').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.nl-NL').click();
-      cy.get('#title_multiloc').type(topicTitle);
-      cy.get('.fr-BE').click();
       cy.get('#title_multiloc').type(topicTitle);
       cy.get('#e2e-submit-wrapper-button').click();
       cy.wait(1000);
